feat(aws-appconfig): move classes & fix comments
