+++ spring-cloud-aws-appconfig/src/test/resources/config.json
