--- spring-cloud-starter-aws-appconfig-config/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-org.springframework.cloud.aws.autoconfigure.appconfig.AwsAppConfigBootstrapConfiguration
